requirements.txt
qiskit>=1.0
mqt.qcec>=2.0
